+++ whatsnew.md
@@ -1,4 +1,3 @@
 # What's new
 
-- On smaller screens you could toggle "launch adjacent" on and off; this has been confusing so the logic has been removed
-- In the app chooser, the letters of the alphabet are highlighted for better visibility
+- Added a settings sheet